pyproject.toml
[project]
name = "electro"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Alexander Kamyanskiy", email = "160958027+kammalexx@users.noreply.github.com"},
]
dependencies = [
    "fastapi>=0.109.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "uvicorn>=0.27.0",
    "pydantic-settings>=2.12.0",
    "dependency-injector>=4.41.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "greenlet>=3.3.1",
    "email-validator>=2.3.0",
    "bcrypt<5.0.0,>=4.1.0",
    "openpyxl>=3.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
create-superuser = {cmd = "python -m backend.cli create-superuser", working_dir = "."}

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
